FixChain/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
chromadb==0.5.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-multipart==0.0.12
numpy==1.26.4
requests==2.32.3
pymongo==4.6.0
motor==3.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==7.4.3
bson==0.5.10